// src/app/Dto/add-car.dto.ts
export interface AddCarDto{
    stanje: string;
    marka: string;
    model: string;
    godiste: number;
    kilometraza: number;
    karoserija: string;
    gorivo: string;
    kubikaza: number;
    snagaMotora: number;
    cena: number;
    fiksnaCena: string;
    zamena: string;
    userId: number;
    images: string[];
}